requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg[binary]
alembic
pydantic-settings
httpx
python-dateutil
pytz
numpy
pandas
PyYAML

pytest
pytest-asyncio
respx


